refactor: restructure MCP architecture into 4 consolidated tools
- Remove Perplexity and Jina (Reader + Grounding) providers
- Consolidate 9-10 tools into 4 outcome-oriented tools:
  web_search, ai_search, github_search, web_extract
- Eliminate unified provider dispatch layer (5 files, 306 LOC)
- Split monolithic utils.ts (800 LOC) into focused modules:
  retry, validation, errors, search_operators, results
- Remove 6 dead functions from utils
- Remove EnhancementProvider/EnhancementResult interfaces
- Move kagi_enrichment into web_search (fix hardcoded query bug)
- Add MCP tool annotations (readOnlyHint, idempotentHint, etc.)
- Per-tool file structure in src/server/tools/
