+++ docker-compose.yml
@@ -13,8 +13,6 @@ services:
     #   - BRAVE_API_KEY=${BRAVE_API_KEY}
     #   - TAVILY_API_KEY=${TAVILY_API_KEY}
     #   - KAGI_API_KEY=${KAGI_API_KEY}
-    #   - PERPLEXITY_API_KEY=${PERPLEXITY_API_KEY}
-    #   - JINA_AI_API_KEY=${JINA_AI_API_KEY}
     #   - FIRECRAWL_API_KEY=${FIRECRAWL_API_KEY}
 
     # Expose port for MCPO
