+++ package.json
@@ -35,8 +35,6 @@
 		"tavily",
 		"brave-search",
 		"kagi",
-		"perplexity",
-		"jina-ai",
 		"firecrawl",
 		"llm",
 		"content-processing",
