+++ src/common/errors.ts
@@ -1,0 +1,49 @@
+import { ErrorType, ProviderError } from './types.js';
+
+export const handle_rate_limit = (
+	provider: string,
+	reset_time?: Date,
+): never => {
+	throw new ProviderError(
+		ErrorType.RATE_LIMIT,
+		`Rate limit exceeded for ${provider}${
+			reset_time ? `. Reset at ${reset_time.toISOString()}` : ''
+		}`,
+		provider,
+		{ reset_time },
+	);
+};
+
+export function handle_provider_error(
+	error: unknown,
+	provider_name: string,
+	operation: string = 'operation',
+): never {
+	if (error instanceof ProviderError) {
+		throw error;
+	}
+	throw new ProviderError(
+		ErrorType.API_ERROR,
+		`Failed to ${operation}: ${
+			error instanceof Error ? error.message : 'Unknown error'
+		}`,
+		provider_name,
+	);
+}
+
+export const sanitize_query = (query: string): string => {
+	return query.trim().replace(/[\n\r]+/g, ' ');
+};
+
+export const create_error_response = (
+	error: Error,
+): { error: string } => {
+	if (error instanceof ProviderError) {
+		return {
+			error: `${error.provider} error: ${error.message}`,
+		};
+	}
+	return {
+		error: `Unexpected error: ${error.message}`,
+	};
+};
