+++ src/common/validation.ts
@@ -1,0 +1,59 @@
+import { ErrorType, ProviderError } from './types.js';
+
+const normalize_api_key = (raw: string): string => {
+	const trimmed = raw.trim();
+	return trimmed.replace(/^(['"])(.*)\1$/, '$2');
+};
+
+export const validate_api_key = (
+	key: string | undefined,
+	provider: string,
+): string => {
+	if (!key) {
+		throw new ProviderError(
+			ErrorType.INVALID_INPUT,
+			`API key not found for ${provider}`,
+			provider,
+		);
+	}
+	return normalize_api_key(key);
+};
+
+export const is_api_key_valid = (
+	key: string | undefined,
+	provider: string,
+): boolean => {
+	if (!key || key.trim() === '') {
+		console.warn(`API key not found or empty for ${provider}`);
+		return false;
+	}
+	return true;
+};
+
+export const is_valid_url = (url: string): boolean => {
+	try {
+		new URL(url);
+		return true;
+	} catch {
+		return false;
+	}
+};
+
+export const validate_processing_urls = (
+	url: string | string[],
+	provider_name: string,
+): string[] => {
+	const urls = Array.isArray(url) ? url : [url];
+
+	for (const u of urls) {
+		if (!is_valid_url(u)) {
+			throw new ProviderError(
+				ErrorType.INVALID_INPUT,
+				`Invalid URL provided: ${u}`,
+				provider_name,
+			);
+		}
+	}
+
+	return urls;
+};
