+++ src/providers/processing/exa_contents/index.ts
@@ -1,15 +1,13 @@
+import { handle_provider_error } from '../../../common/errors.js';
 import { http_json } from '../../../common/http.js';
+import { retry_with_backoff } from '../../../common/retry.js';
 import {
 	ErrorType,
 	ProcessingProvider,
 	ProcessingResult,
 	ProviderError,
 } from '../../../common/types.js';
-import {
-	handle_provider_error,
-	retry_with_backoff,
-	validate_api_key,
-} from '../../../common/utils.js';
+import { validate_api_key } from '../../../common/validation.js';
 import { config } from '../../../config/env.js';
 
 interface ExaContentsRequest {
