+++ src/providers/processing/firecrawl_extract/index.ts
@@ -1,9 +1,10 @@
+import { handle_provider_error } from '../../../common/errors.js';
 import {
 	make_firecrawl_request,
 	poll_firecrawl_job,
 	validate_firecrawl_response,
-	validate_firecrawl_urls,
 } from '../../../common/firecrawl_utils.js';
+import { retry_with_backoff } from '../../../common/retry.js';
 import {
 	ErrorType,
 	ProcessingProvider,
@@ -11,10 +12,9 @@ import {
 	ProviderError,
 } from '../../../common/types.js';
 import {
-	handle_provider_error,
-	retry_with_backoff,
 	validate_api_key,
-} from '../../../common/utils.js';
+	validate_processing_urls,
+} from '../../../common/validation.js';
 import { config } from '../../../config/env.js';
 
 interface FirecrawlExtractResponse {
@@ -41,7 +41,7 @@ export class FirecrawlExtractProvider implements ProcessingProvider {
 		extract_depth: 'basic' | 'advanced' = 'basic',
 	): Promise<ProcessingResult> {
 		// Extract works with a single URL at a time
-		const urls = validate_firecrawl_urls(url, this.name);
+		const urls = validate_processing_urls(url, this.name);
 		const extract_url = urls[0];
 
 		const extract_request = async () => {
