+++ src/providers/search/brave/index.ts
@@ -1,17 +1,17 @@
+import { handle_provider_error } from '../../../common/errors.js';
 import { http_json } from '../../../common/http.js';
+import { retry_with_backoff } from '../../../common/retry.js';
+import {
+	apply_search_operators,
+	build_query_with_operators,
+	parse_search_operators,
+} from '../../../common/search_operators.js';
 import {
 	BaseSearchParams,
 	SearchProvider,
 	SearchResult,
 } from '../../../common/types.js';
-import {
-	apply_search_operators,
-	build_query_with_operators,
-	handle_provider_error,
-	parse_search_operators,
-	retry_with_backoff,
-	validate_api_key,
-} from '../../../common/utils.js';
+import { validate_api_key } from '../../../common/validation.js';
 import { config } from '../../../config/env.js';
 
 interface BraveSearchResponse {
